fix: fixed circular dependencies
