+++ packages/@webex/internal-plugin-metrics/package.json
@@ -26,7 +26,6 @@
   "dependencies": {
     "@webex/common": "workspace:^",
     "@webex/common-timers": "workspace:^",
-    "@webex/internal-plugin-device": "workspace:^",
     "@webex/internal-plugin-metrics": "workspace:^",
     "@webex/test-helper-chai": "workspace:^",
     "@webex/test-helper-mock-webex": "workspace:^",
