+++ packages/@webex/internal-plugin-metrics/src/index.js
@@ -2,8 +2,6 @@
  * Copyright (c) 2015-2020 Cisco Systems, Inc. See LICENSE file.
  */
 
-import '@webex/internal-plugin-device';
-
 import {registerInternalPlugin} from '@webex/webex-core';
 
 import Metrics from './metrics';
