+++ packages/@webex/internal-plugin-metrics/src/metrics.js
@@ -70,7 +70,6 @@ const Metrics = WebexPlugin.extend({
       domain:
         typeof window !== 'undefined' ? window.location.hostname || 'non-browser' : 'non-browser', // Check what else we could measure
       client_id: this.webex.credentials.config.client_id,
-      user_id: this.webex.internal.device.userId,
     };
 
     try {
