+++ packages/@webex/internal-plugin-metrics/test/unit/spec/metrics.js
@@ -89,7 +89,7 @@ describe('plugin-metrics', () => {
       webex.credentials = new Credentials(undefined, {parent: webex});
       sinon.stub(webex.credentials, 'getClientToken').returns(Promise.resolve('token'));
 
-      webex.internal = {...webex.internal, device: {userId: 'userId'}};
+      webex.internal = {...webex.internal};
       webex.config = {
         ...webex.config,
         appName: 'appName',
