+++ packages/@webex/webex-core/src/lib/services/interceptors/server-error.js
@@ -3,7 +3,7 @@
  */
 
 import {Interceptor} from '@webex/http-core';
-import {WebexHttpError} from '@webex/webex-core';
+import WebexHttpError from '../../webex-http-error';
 /**
  * Changes server url when it fails
  */
